requirements.txt
# Core dependencies
torch>=2.0.0
torch_geometric>=2.4.0

# Optional PyG dependencies (install if needed for sparse operations)
# torch-scatter
# torch-sparse

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0

# Visualization and analysis
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# Experiment tracking and utilities
tqdm>=4.65.0
pyyaml>=6.0

# Testing
pytest>=7.3.0

# Jupyter for notebooks
jupyter>=1.0.0
ipykernel>=6.22.0
